refactor: extract DatasetStatsRow; move formatCompactNumber to lib
