+++ src/lib/dataset-stats.ts
@@ -22,6 +22,16 @@ export function processDatasetStats(dataset: Pick<Dataset, 'dataCount' | 'stats'
   };
 }
 
+/**
+ * Format number in compact notation (1.2k, 2M, etc.)
+ */
+export function formatCompactNumber(value: number | string): string {
+  const num = typeof value === 'string' ? parseInt(value.replace(/,/g, ''), 10) : value;
+  if (num < 1000) return num.toString();
+  if (num < 1000000) return `${(num / 1000).toFixed(1).replace(/\.0$/, '')}k`;
+  return `${(num / 1000000).toFixed(1).replace(/\.0$/, '')}M`;
+}
+
 export function formatRelativeTime(timestamp: string | Date | null | undefined, locale: string): string {
   if (!timestamp) return "—";
